+++ sat/model/__init__.py
@@ -1,4 +1,5 @@
-from .base_model import BaseModel, BaseMixin, AutoModel
+from .base_model import BaseModel, BaseMixin, AutoModel, get_model
 from .cached_autoregressive_model import CachedAutoregressiveModel
 from .encoder_decoder_model import EncoderDecoderModel
 from .official import *
+