+++ sat/mpu/initialize.py
@@ -19,6 +19,7 @@
 import torch
 
 from .utils import ensure_divisibility
+from sat.helpers import print_rank0
 
 
 # Model parallel group that the current rank belongs to.
@@ -47,7 +48,7 @@ def initialize_model_parallel(model_parallel_size_):
     ranks 8 to 15 belong to the second box.
     """
     if torch.distributed.get_rank() == 0:
-        print('> initializing model parallel with size {}'.format(
+        print_rank0('> initializing model parallel with size {}'.format(
             model_parallel_size_))
     # Get world size and rank. Ensure some consistencies.
     assert torch.distributed.is_initialized()
@@ -110,7 +111,7 @@ def get_model_parallel_rank():
 
 
 def get_model_parallel_src_rank():
-    """Calculate the global rank corresponding to a local rank zeor
+    """Calculate the global rank corresponding to a local rank zero
     in the model parallel group."""
     global_rank = torch.distributed.get_rank()
     local_world_size = get_model_parallel_world_size()